// mono-repo/packages/gantry-app/src/app/cds/matches/snooker-cds/models/snooker-cds-template.model.ts
import { SportContentParameters } from '../../../../common/models/sport-content/sport-content-parameters.model';

export class SnookerCdsTemplateResult {
    games?: Game[];
    eventStartDate?: string;
    sportName?: string;
    title?: string;
    competitionName: string;
    content: SnookerContentParams;
    frameBetting: CorrectScorer;
    context: string;
    marketVersesName?: string;
    homeName: string;
    awayName: string;
    drawTitle?: string;
}

export interface Game {
    id?: number;
    gameName?: string;
    marketName?: string;
    matchBetting?: Selection;
    correctScore?: CorrectScorer;
    totalFrames?: Selection;
    matchHanicap?: Selection;
    isMatchBetting?: boolean;
    isFrameBetting?: boolean;
    isHomeScorer?: boolean;
    isAwayScorer?: boolean;
    isTotalFrames?: boolean;
    isMatchHandicap?: boolean;
}

export interface Selection {
    homePrice?: string;
    homePlayer?: string;
    awayPrice?: string;
    awayPlayer?: string;
    drawPrice?: string;
    drawPlayer?: string;
    drawSuspended?: boolean;
}

export class CorrectScorer {
    homeTeamScorerList?: Array<BetDetails>;
    awayTeamScorerList?: Array<BetDetails>;
}

export class SnookerContentParams extends SportContentParameters {}

export class BetDetails {
    betName: string;
    betOdds: string;
}